+++ CHANGELOG.md
@@ -102,6 +102,10 @@
 
   ([Igor Castejón](https://github.com/IgorCastejon))
 
+- The language server now performs best-effort zero value generation for
+  `decode.failure` when using the `Generate Dynamic Decoder` code action.
+  ([Hari Mohan](https://github.com/seafoamteal))
+
 ### Formatter
 
 ### Bug fixes